holidays v0.0.2 Holidays.DateCalculator.Easter

Functions

gregorian_easter_for(year)

Returns the date of Easter for the given year

Examples

iex> Holidays.DateCalculator.Easter.gregorian_easter_for(2016)
{2016, 3, 27}

iex> Holidays.DateCalculator.Easter.gregorian_easter_for(2015)
{2015, 4, 5}

gregorian_orthodox_easter_for(year)

Returns the date of Orthodox Easter in the given year

Examples

iex> Holidays.DateCalculator.Easter.gregorian_orthodox_easter_for(2016)
{2016, 5, 1}

iex> Holidays.DateCalculator.Easter.gregorian_orthodox_easter_for(1815)
{1815, 4, 30}

iex> Holidays.DateCalculator.Easter.gregorian_orthodox_easter_for(2101)
{2101, 4, 24}

julian_orthodox_easter_for(year)